User service: remove wrapper
